+++ test/components/folio/console/files/show/file_placements_component_test.rb
@@ -20,4 +20,80 @@ class Folio::Console::Files::Show::FilePlacementsComponentTest < Folio::Console:
     assert_selector(".f-c-files-show-file-placements")
     assert_selector(".f-c-files-show-file-placements__table")
   end
+
+  def test_render_orphaned_placement
+    video = create(:folio_file_video)
+    Folio::FilePlacement::VideoCover.create!(placement: nil, file: video)
+
+    render_inline(Folio::Console::Files::Show::FilePlacementsComponent.new(file: video))
+
+    assert_selector(".f-c-files-show-file-placements__row--orphaned")
+    assert_selector(".f-c-files-show-file-placements__row--orphaned",
+                    text: I18n.t("folio.console.files.show.file_placements_component.orphaned"))
+  end
+
+  def test_render_destroy_button_for_orphaned_placement_only
+    video = create(:folio_file_video)
+    page = create(:folio_page, published: true)
+    orphaned = Folio::FilePlacement::VideoCover.create!(placement: nil, file: video)
+    Folio::FilePlacement::VideoCover.create!(placement: page, file: video)
+
+    render_inline(Folio::Console::Files::Show::FilePlacementsComponent.new(file: video))
+
+    assert_selector(".f-c-files-show-file-placements__destroy-cell a[data-method='delete']",
+                    text: I18n.t("folio.console.files.show.file_placements_component.destroy_orphan"),
+                    count: 1)
+    assert_selector(".f-c-files-show-file-placements__row--orphaned " \
+                    ".f-c-files-show-file-placements__destroy-cell " \
+                    "a[href$='/console/file_placements/#{orphaned.id}']")
+    assert_no_selector(".f-c-files-show-file-placements__row:not(.f-c-files-show-file-placements__row--orphaned) " \
+                       "a[data-method='delete']")
+  end
+
+  def test_render_orphaned_placement_with_title_snapshot
+    video = create(:folio_file_video)
+    Folio::FilePlacement::VideoCover.create!(placement: nil,
+                                             file: video,
+                                             placement_title: "Folio::Page - Smazaná stránka",
+                                             placement_title_type: "Folio::Page")
+
+    render_inline(Folio::Console::Files::Show::FilePlacementsComponent.new(file: video))
+
+    assert_selector(".f-c-files-show-file-placements__row--orphaned",
+                    text: "Folio::Page - Smazaná stránka")
+    assert_selector(".f-c-files-show-file-placements__row--orphaned",
+                    text: Folio::Page.model_name.human)
+  end
+
+  def test_render_published_owner_state
+    video = create(:folio_file_video)
+    page = create(:folio_page, published: true)
+    Folio::FilePlacement::VideoCover.create!(placement: page, file: video)
+
+    render_inline(Folio::Console::Files::Show::FilePlacementsComponent.new(file: video))
+
+    assert_selector(".f-c-files-show-file-placements__row .text-success",
+                    text: I18n.t("folio.console.files.show.file_placements_component.published"))
+    assert_no_selector(".f-c-files-show-file-placements__row--orphaned")
+  end
+
+  def test_render_unpublished_owner_state
+    video = create(:folio_file_video)
+    page = create(:folio_page, published: false)
+    Folio::FilePlacement::VideoCover.create!(placement: page, file: video)
+
+    render_inline(Folio::Console::Files::Show::FilePlacementsComponent.new(file: video))
+
+    assert_selector(".f-c-files-show-file-placements__row .text-danger",
+                    text: I18n.t("folio.console.files.show.file_placements_component.unpublished"))
+  end
+
+  def test_render_owner_label_for_regular_placement
+    file_placement = create(:folio_file_placement_cover)
+
+    render_inline(Folio::Console::Files::Show::FilePlacementsComponent.new(file: file_placement.file))
+
+    assert_selector(".f-c-files-show-file-placements__row",
+                    text: file_placement.placement.to_label)
+  end
 end